+++ Frameworks/Lablib/LLPlotAxes.m
@@ -229,7 +229,7 @@
 // Return the number of digits of precision that the tick labels should 
 // be displayed with.
 
-+ (long) precisionForMin:(float)axisMinValue andMax:(float)axisMaxValue {
++ (int) precisionForMin:(float)axisMinValue andMax:(float)axisMaxValue {
 
     long precision;
     float value;
